+++ http4k-contract/src/main/kotlin/org/http4k/contract/security/OpenIdConnectSecurity.kt
@@ -1,0 +1,12 @@
+package org.http4k.contract.security
+
+import org.http4k.core.Filter
+import org.http4k.core.Uri
+
+class OpenIdConnectSecurity(
+    val discoveryUrl: Uri,
+    override val filter: Filter,
+    val name: String = "openIdConnect"
+): Security {
+    companion object
+}