+++ http4k-contract/src/test/kotlin/org/http4k/contract/ui/SwaggerUiTest.kt
@@ -3,12 +3,13 @@ package org.http4k.contract.ui
 import org.http4k.core.Method.GET
 import org.http4k.core.Request
 import org.http4k.core.Uri
+import org.http4k.testing.ApprovalTest
 import org.http4k.testing.Approver
 import org.http4k.testing.HtmlApprovalTest
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 
-@ExtendWith(HtmlApprovalTest::class)
+@ExtendWith(ApprovalTest::class)
 class SwaggerUiTest {
 
     @Test
@@ -20,5 +21,16 @@ class SwaggerUiTest {
             requestSnippetsEnabled = true
         )
         approver.assertApproved(handler(Request(GET, "")))
+    }
+
+    @Test
+    fun `can serve swagger oauth2 redirect`(approver: Approver) {
+        val handler = swaggerUi(
+            Uri.of("/spec"),
+            title = "Cat Shelter",
+            displayOperationId = true,
+            requestSnippetsEnabled = true
+        )
+        approver.assertApproved(handler(Request(GET, "oauth2-redirect.html")))
     }
 }